# Ed: match stop_area and poi by using admin boundary if possible