When OSM data is used we git the boundary of each admins, in that case
we want to use these boundary to associate stop_points to an admin.

This doesn't use any kind of spacial indexes, so it might be slower than
before.

If we do not find any admin by using the boundary we still keep the old
approach as a fallback, so if there is no boundaries it work as before.


|coverage|before|before with cities|after|after with cities|
|---|-------------|---------------------|---------------------|----------------------|
|bordeaux| 7042ms| 6974ms | 417ms | 287ms |
|fr-idf| 70811ms | 50539ms | 153190ms| 118166ms|

Bordeaux has only a few admins, so the after result is very fast, also, no call to cities is done, so it has no effect on perf.
So funnily cities make the admin search faster...