@TeXitoi 2018-06-26
The "only in cities" case is much better with cities:
 - without cities, each times, search all admins, then old fashion admin search
 - withi cities, first time search all admins, then old fashion admin search, then cities, and the others, direct found in all admins.

This might explain the performance boost.